interview-prep/requirements.txt
google-genai
opencv-python
pyaudio
pillow
mss
dotenv


openai-whisper
torch
pyaudio
webrtcvad-wheels 